Refactor store project to work with modules
