+++ src/modules/presidents/store/index.js
@@ -1,0 +1,10 @@
+import state from './state';
+import mutations from './mutations';
+import actions from './actions';
+
+export default {
+  namespaced: true,
+  state,
+  mutations,
+  actions,
+};
