+++ src/modules/presidents/store/state.js
@@ -1,0 +1,5 @@
+const state = {
+  candidatos: [],
+};
+
+export default state;
