+++ src/store/modules.js
@@ -1,0 +1,7 @@
+import presidents from '@/modules/presidents/store';
+
+const modules = {
+  presidents,
+};
+
+export default modules;
